+++ example/lib/env_test2.g.dart
@@ -1,6 +1,6 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
-part of 'env2.dart';
+part of 'env_test2.dart';
 
 // **************************************************************************
 // EnvysticGenerator
@@ -8,31 +8,19 @@ part of 'env2.dart';
 
 const String? _encryptionKey = null;
 const String _encodedEntries =
-    'eyJ0ZXN0U3RyaW5nIjoiZEdWemRGTjBjbWx1Wnc9PSIsInRlc3RJbnQiOiJNVEl6IiwidGVzdERvdWJsZSI6Ik1TNHlNdz09IiwidGVzdEJvb2wiOiJkSEoxWlE9PSIsInRlc3REeW5hbWljIjoiTVRJeiJ9';
+    'eyJ0ZXN0U3RyaW5nIjpudWxsLCJ0ZXN0SW50IjpudWxsfQ==';
 
-class _$Env2 extends Env2 implements EnvysticInterface {
-  const _$Env2() : super._();
+class _$EnvTest2 extends EnvTest2 implements EnvysticInterface {
+  const _$EnvTest2() : super._();
   @override
   String get pairKeyEncodedEntries$ => _encodedEntries;
 
   @override
-  String get testString =>
+  String? get testString =>
       getEntryValue('testString', _encodedEntries, _encryptionKey);
 
   @override
-  int get testInt => getEntryValue('testInt', _encodedEntries, _encryptionKey);
-
-  @override
-  double get testDouble =>
-      getEntryValue('testDouble', _encodedEntries, _encryptionKey);
-
-  @override
-  bool get testBool =>
-      getEntryValue('testBool', _encodedEntries, _encryptionKey);
-
-  @override
-  dynamic get testDynamic =>
-      getEntryValue('testDynamic', _encodedEntries, _encryptionKey);
+  int? get testInt => getEntryValue('testInt', _encodedEntries, _encryptionKey);
 
   @override
   bool operator ==(Object other) =>