+++ packages/envystic/lib/envystic.dart
@@ -3,4 +3,5 @@ library envystic;
 export 'src/annotation/envystic.dart';
 export 'src/annotation/envystic_field.dart';
 export 'src/annotation/key_format.dart';
-export 'src/utils.dart';
+export 'src/envystic_interface.dart';
+export 'src/helper/utils.dart';